requirements.txt
# Use Microsoft's ORT nightly PyPI feed for onnxruntime-genai if needed.
# When installing with pip directly you can run:
#   pip install --index-url https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/ORT-Nightly/pypi/simple/ onnxruntime-genai
# Or simply install the full requirements below; pip will use the extra index URL for that package.
--extra-index-url https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/ORT-Nightly/pypi/simple/

huggingface_hub==0.35.3
streamlit==1.50.0
PyYAML==6.0.2
torch==2.5.1
transformers>=4.40.0
onnx>=1.16.0
onnxruntime-genai
